testing_models/huggingface/requirements.txt
gradio
hopsworks
joblib
pandas
python-dotenv
google-api-python-client
